Redux: Restructure redux components
