+++ template/src/presentation/shared-state/redux/actions/authentication.action.ts
