+++ template/src/presentation/shared-state/redux/actions/configuration.action.ts
