+++ template/src/presentation/shared-state/redux/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './authentication.action';
+export * from './configuration.action';
