--- template/src/presentation/shared-state/redux/authentication/index.ts
@@ -1,4 +1,0 @@
-export * from './type';
-export * from './action';
-export * from './reducer';
-export * from './epic';
