--- template/src/presentation/shared-state/redux/authentication/type.ts
@@ -1,5 +1,0 @@
-export type AuthenticationState = {
-  isAuthorized: boolean;
-  isAuthenticating: boolean;
-  isAuthenticatingLocally: boolean;
-};
