--- template/src/presentation/shared-state/redux/configuration/type.ts
@@ -1,5 +1,0 @@
-import {ThemeConfig} from '@core';
-
-export type ConfigurationState = {
-  themeConfig: ThemeConfig;
-};
