+++ template/src/presentation/shared-state/redux/epics/authentication.epic.ts
@@ -14,7 +14,7 @@ import {
   signInLocally,
   signInLocallyFailed,
   signInLocallySuccess,
-} from './action';
+} from '../actions';
 
 import {SignInUseCase} from '@domain';
