+++ template/src/presentation/shared-state/redux/epics/configuration.epic.ts
@@ -2,7 +2,7 @@ import {Epic, combineEpics} from 'redux-observable';
 import {of} from 'rxjs';
 import {filter, mergeMap} from 'rxjs/operators';
 
-import {ConfigurationEpicActions, setTheme, setThemeSuccess} from './action';
+import {ConfigurationEpicActions, setTheme, setThemeSuccess} from '../actions';
 
 const setThemeEpic$: Epic<ConfigurationEpicActions> = (action$) =>
   action$.pipe(
