+++ template/src/presentation/shared-state/redux/epics/index.ts
@@ -1,0 +1,2 @@
+export * from './authentication.epic';
+export * from './configuration.epic';
