+++ template/src/presentation/shared-state/redux/index.ts
@@ -1,4 +1,3 @@
 export * from './store';
 export * from './types';
-export * from './authentication';
-export * from './configuration';
+export * from './actions';
