+++ template/src/presentation/shared-state/redux/reducers/index.ts
@@ -1,0 +1,2 @@
+export * from './authentication.reducer';
+export * from './configuration.reducer';
